ai/requirements.txt
langchain_community==0.4.1
langchain_huggingface==1.2.1
langchain_openai==1.1.12
langchain_text_splitters==1.1.1
langgraph==1.1.6
protobuf==7.34.1
python-dotenv==1.2.2
Requests==2.33.1
